refactor: Introduce normalization for whitespace handling and restructure lexer modules
- Added a new `normalize_whitespace` module to handle whitespace remainder transformations.
- Updated the lexer to use the new `pipeline` module for core tokenization.
- Renamed and reorganized existing transformation references to improve clarity and maintainability.
- Ensured that whitespace processing adheres to the specified behavior for indentation and text handling.
